knowledge-vault/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.4.2
gunicorn>=21.2.0

# Firebase & Google Cloud
firebase-admin>=6.2.0

# OpenAI for embeddings
openai>=1.0.0
tenacity>=8.2.3

# Document processing
pypdf>=3.17.1
docx2txt>=0.8
beautifulsoup4>=4.12.2
